modules/api-locations-consumer/requirements.txt
kafka-python == 2.0.2
psycopg2-binary==2.8.5
SQLAlchemy==1.3.19
